Move MultipleIPInput to src/components
