+++ packages/manager/src/components/MultipleIPInput/MultipleIPInput.test.tsx
@@ -1,0 +1,9 @@
+// import * as React from 'react';
+
+// import MultipleIPInput from './MultipleIPInput';
+
+describe('MultipleIPInput', () => {
+  it.skip('should render a single input field when ips is an empty array', () => {
+    expect(true).toBe(true);
+  });
+});
