+++ packages/manager/src/components/MultipleIPInput/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './MultipleIPInput';
